requirements.txt
flask>=2.3.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
werkzeug>=2.3.0
tensorflow>=2.13.0
ultralytics>=8.0.0
clarifai>=10.0.0
easyocr>=1.7.0
google-genai>=1.0.0
scikit-learn>=1.3.0
